add peerSetManager + remove deprec. method
